requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
annotated-types==0.6.0
anyio==4.0.0
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.2
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
idna==3.4
iniconfig==2.0.0
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
packaging==23.2
pluggy==1.3.0
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic_core==2.14.1
pytest==7.4.3
python-dotenv==1.0.0
referencing==0.30.2
requests==2.31.0
rpds-py==0.12.0
sniffio==1.3.0
SQLAlchemy==2.0.23
typing_extensions==4.8.0
urllib3==2.1.0
